ml-model-creation/requirements.txt
numpy
scipy
Flask
flask-restplus
pandas
scikit_learn
joblib
requests
sklearn-pmml-model
